Item speed is now non const
Now item speed is changed each time move_to_position is called.
Set up different speeds for when item is dragged around and put in inventory, for a smoother transition when putting item in inventory.
Also added a default item speed so that speed argument to move_to_position is not required.